refactor: 강의 신청 관련 로직 커스텀 훅으로 분리
