+++ src/app/(main)/lecture/[id]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect, useCallback } from "react";
+import { useState, useEffect } from "react";
 import { notFound, useParams } from "next/navigation";
 import LectureImage from "@/components/LectureImage";
 import Image from "next/image";
@@ -8,10 +8,7 @@ import { lecturesApi } from "@/app/api/lectures";
 import { LectureSidebar } from "@/components/LectureSidebar";
 import { LectureWithCoach } from "@/types/lectures";
 import { MobileFloatingBar } from "@/components/MobileFloatingBar";
-import InfoModal from "@/components/InfoModal";
-import { User } from "@supabase/supabase-js";
-import { supabase } from "@/utils/supabase";
-import { leadsApi } from "@/app/api/leads";
+import { useLectureApply } from "@/hooks/useLectureApply";
 
 export default function CourseDetailPage() {
   const [activeTab, setActiveTab] = useState("intro");
@@ -19,88 +16,9 @@ export default function CourseDetailPage() {
   const lectureId = params.id as string;
   const [lecture, setLecture] = useState<LectureWithCoach | null>(null);
   const [loading, setLoading] = useState(true);
-  const [modalStatus, setModalStatus] = useState<
-    "hidden" | "login" | "success"
-  >("hidden");
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [user, setUser] = useState<User | null>(null);
-  const [applyAfterLogin, setApplyAfterLogin] = useState(false);
 
-  const handleLogin = async () => {
-    const { error } = await supabase.auth.signInWithOAuth({
-      provider: "kakao",
-      options: {
-        redirectTo: `${window.location.origin}${window.location.pathname}${window.location.search}`,
-      },
-    });
-
-    if (error) {
-      console.error("Kakao login error:", error);
-      alert("로그인 중 오류가 발생했습니다. 다시 시도해 주세요.");
-    }
-  };
-
-  const handleApply = useCallback(
-    async (currentUser: User) => {
-      if (!lecture) return;
-
-      setIsLoading(true);
-      setError(null);
-
-      try {
-        await leadsApi.createLead({
-          name: currentUser.user_metadata?.full_name || currentUser.email || "",
-          email: currentUser.email || "",
-          subscribe: lecture.id,
-        });
-        setModalStatus("success");
-      } catch (err) {
-        if (err instanceof Error) {
-          setError(err.message);
-        } else {
-          setError("알 수 없는 오류가 발생하여 신청에 실패했습니다.");
-        }
-      } finally {
-        setIsLoading(false);
-      }
-    },
-    [lecture],
-  );
-
-  const handleApplyClick = () => {
-    if (!user) {
-      setApplyAfterLogin(true);
-      setModalStatus("login");
-      return;
-    }
-    handleApply(user);
-  };
-
-  useEffect(() => {
-    if (error) {
-      alert(error);
-      setError(null);
-    }
-  }, [error]);
-
-  useEffect(() => {
-    const {
-      data: { subscription },
-    } = supabase.auth.onAuthStateChange((event, session) => {
-      setUser(session?.user ?? null);
-
-      if (event === "SIGNED_IN" && session && applyAfterLogin) {
-        setApplyAfterLogin(false);
-        setModalStatus("hidden");
-        handleApply(session.user);
-      }
-    });
-
-    return () => {
-      subscription.unsubscribe();
-    };
-  }, [applyAfterLogin, handleApply]);
+  // 커스텀 훅 사용
+  const { isLoading, handleApplyClick } = useLectureApply(lecture);
 
   useEffect(() => {
     const fetchLecture = async () => {
@@ -236,13 +154,6 @@ export default function CourseDetailPage() {
         lecture={lecture}
         onApplyClick={handleApplyClick}
         isLoading={isLoading}
-      />
-
-      <InfoModal
-        isOpen={modalStatus !== "hidden"}
-        onClose={() => setModalStatus("hidden")}
-        onLogin={handleLogin}
-        status={modalStatus === "success" ? "success" : "login"}
       />
     </>
   );
